pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "spot_planning_demo"
version = "0.1.0"
description = "Bilevel planning demo with Spot robot."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
   "gymnasium==1.2.0",
   "numpy>=1.23.5,<2.0",
   "pybullet==3.2.6",
   "pybullet_helpers@git+https://github.com/tomsilver/pybullet-helpers.git",
   "opencv-python == 4.7.0.72",
   "scipy",
   "types-requests",
   "bosdyn-client",
]

[project.optional-dependencies]
develop = [
    "black",
    "docformatter",
    "isort",
    "mypy",
    "pylint>=2.14.5",
    "pytest-pylint>=0.18.0",
    "pytest>=7.2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
spot_planning_demo = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
py_version = 311
profile = "black"
multi_line_output = 2
skip_glob = ["venv/*", ".venv/*"]
split_on_trailing_comma = true

[tool.docformatter]
line-length = 88
wrap-summaries = 88
wrap-descriptions = 88

[tool.mypy]
strict_equality = true
disallow_untyped_calls = true
warn_unreachable = true
exclude = ["venv/*", ".venv/*"]

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "pybullet.*",
    "bosdyn.*",
    "cv2.*",
    "scipy.*",
    "google.*",
    "dill.*",
]
ignore_missing_imports = true
